moved non int test to non int test file
